requirements.txt
aiozipstream==0.4
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
boto3==1.26.138
botocore==1.29.138
brevitas==0.8.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
comm==0.1.3
concrete-ml==1.0.2
concrete-python==1.0.0
customtkinter==5.1.3
darkdetect==0.8.0
debugpy==1.6.7
decorator==5.1.1
dependencies==2.0.1
dill==0.3.6
Django==4.2.1
executing==1.2.0
fastapi==0.93.0
filelock==3.12.0
flatbuffers==23.5.9
fsspec==2023.5.0
h11==0.14.0
huggingface-hub==0.14.1
humanfriendly==10.0
hummingbird-ml==0.4.8
idna==3.4
image==1.5.33
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.13.1
onnxconverter-common==1.13.0
onnxmltools==1.11.0
onnxoptimizer==0.3.10
onnxruntime==1.13.1
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.7
Pygments==2.15.1
pygubu==0.31
pygubu-designer==0.36
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.1.3
scipy==1.10.1
screeninfo==0.8.1
six==1.16.0
skl2onnx==1.12
skops==0.5.0
skorch==0.11.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.25.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.21.1
wcwidth==0.2.6
xgboost==1.6.2
